phase-1-foundations/requirements.txt
chromadb>=0.4.0
langchain>=0.3.0
langchain-together>=0.2.0
python-dotenv>=1.0.0
